Move all MCP tests into mcp.test.ts + remove custom build commands
