+++ package.json
@@ -18,8 +18,6 @@
     "build:sdk-docs": "sh scripts/jsdoc.sh",
     "build:create-leo-app": "cd create-leo-app && yarn build",
     "build:all": "yarn build:wasm && yarn build:sdk && yarn build:create-leo-app",
-    "build:wasm-debug": "rustup show active-toolchain && cd wasm && CARGO_PROFILE_TEST_LTO=off BUILD_DEBUG=1 yarn build",
-    "build:all-debug": "yarn build:wasm-debug && yarn build:sdk && yarn build:create-leo-app",
     "start:website": "cd website && yarn dev",
     "test:wasm": "cd wasm && yarn test",
     "test:sdk": "cd sdk && yarn test",
