+++ sdk/tests/program-manager.test.ts
@@ -10,7 +10,6 @@ import {
     ImportedPrograms,
     ImportedVerifyingKeys,
     OfflineQuery,
-    Plaintext,
     PrivateKey,
     Program,
     ProgramManager,
@@ -19,7 +18,7 @@ import {
     Transaction,
     verifyFunctionExecution,
     VerifyingKey,
-    ViewKey
+    ViewKey,
 } from "@provablehq/sdk/%%NETWORK%%.js";
 import {
     beaconAddressString,
@@ -472,239 +471,4 @@ describe('Program Manager', async () => {
         });
     });
 
-    describe('Program Manager public message signing (computeExternalSigningInputs)', () => {
-        const privateKeyStr = "APrivateKey1zkp7Vc4xJt8HqW9U7VhY6h32d8Z9Xi5C6ZZX3gtXxbBSJmj";
-        const beaconAddress = "aleo1rhgdu77hgyqd3xjj8ucu3jj9r2krwz6mnzyd80gncr5fxcwlh5rsvzp9px";
-        const transferPublicInputs = [
-            beaconAddress,
-            "100u64",
-        ];
-        const transferPublicInputTypes = ["address.public", "u64.public"];
-        // Record owned by beacon address (required for ExecutionRequest.sign - record must belong to signer)
-        const recordBeaconOwned = `{ owner: ${beaconAddress}.private, microcredits: 1000000u64.private, _nonce: 3634848344765318974603121890869676775499130077229666060613233255327643175219group.public, _version: 1u8.public }`;
-        const transferPrivateInputs = [
-            recordBeaconOwned,
-            "aleo1rhgdu77hgyqd3xjj8ucu3jj9r2krwz6mnzyd80gncr5fxcwlh5rsvzp9px",
-            "100u64",
-        ];
-        const transferPrivateInputTypes = ["credits.record", "address.private", "u64.private"];
-
-        it('Should compute external signing inputs and return ExternalSigningInput shape (functionId, isRoot, requestInputs, checksum)', async () => {
-            const externalSigningInputs = await programManager.computeExternalSigningInputs({
-                programName: "credits.aleo",
-                functionName: "transfer_public",
-                inputs: transferPublicInputs,
-                inputTypes: transferPublicInputTypes,
-                isRoot: true,
-                checksum: null,
-            });
-
-            expect(externalSigningInputs).to.be.an("object");
-            expect(externalSigningInputs).to.have.property("functionId");
-            expect(externalSigningInputs).to.have.property("isRoot");
-            expect(externalSigningInputs).to.have.property("requestInputs");
-            expect(externalSigningInputs.functionId).to.be.a("string");
-            expect(externalSigningInputs.functionId.length).to.be.greaterThan(0);
-            expect(externalSigningInputs.functionId).to.match(/field$/);
-            expect(externalSigningInputs.isRoot).to.equal("1field");
-            expect(externalSigningInputs.checksum).to.satisfy((v: unknown) => v === null || v === undefined);
-            expect(externalSigningInputs.requestInputs).to.be.an("array").with.lengthOf(2);
-
-            const [firstInput, secondInput] = externalSigningInputs.requestInputs;
-            expect(firstInput).to.have.property("outputType", "public");
-            expect(firstInput).to.have.property("index", "0field");
-            expect(firstInput).to.have.property("data").that.is.an("array");
-            expect(firstInput.data.length).to.be.greaterThan(0);
-            expect(secondInput).to.have.property("outputType", "public");
-            expect(secondInput).to.have.property("index", "1field");
-            expect(secondInput).to.have.property("data").that.is.an("array");
-
-            // Assert each request input's field representation round-trips through Plaintext (fromFields -> toFields -> string).
-            const assertFieldsRoundTripThroughPlaintext = (fieldStrings: string[], expectedString: string) => {
-                const fields = fieldStrings.map((s: string) => Field.fromString(s));
-                const plaintext = Plaintext.fromFields(fields);
-                const backFields = plaintext.toFields();
-                const backStrings = Array.from(backFields).map((f: Field) => f.toString());
-                expect(backStrings).to.deep.equal(fieldStrings);
-                expect(plaintext.toString()).to.equal(expectedString);
-            };
-            assertFieldsRoundTripThroughPlaintext(firstInput.data, transferPublicInputs[0]);
-            assertFieldsRoundTripThroughPlaintext(secondInput.data, transferPublicInputs[1]);
-        });
-
-        it('Should match ExecutionRequest.sign for transfer_public', async () => {
-            const privateKey = PrivateKey.from_string(privateKeyStr);
-            const viewKey = ViewKey.from_private_key(privateKey);
-            const externalSigningInputs = await programManager.computeExternalSigningInputs({
-                programName: "credits.aleo",
-                functionName: "transfer_public",
-                inputs: transferPublicInputs,
-                inputTypes: transferPublicInputTypes,
-                isRoot: true,
-                checksum: null,
-            });
-
-            const signedRequest = ExecutionRequest.sign(
-                privateKey,
-                "credits.aleo",
-                "transfer_public",
-                transferPublicInputs,
-                transferPublicInputTypes,
-                undefined,
-                undefined,
-                true,
-            );
-            expect(signedRequest.program_id()).to.equal("credits.aleo");
-            expect(signedRequest.function_name()).to.equal("transfer_public");
-            expect(signedRequest.inputs().length).to.equal(2);
-            expect(externalSigningInputs.requestInputs.length).to.equal(signedRequest.inputs().length);
-
-            // Build the same request via fromExternallySignedData using view_key to compute input_ids.
-            const externallySignedRequest = ExecutionRequest.fromExternallySignedData(
-                "credits.aleo",
-                "transfer_public",
-                transferPublicInputs,
-                transferPublicInputTypes,
-                signedRequest.signature(),
-                signedRequest.tvk(),
-                signedRequest.signer(),
-                signedRequest.sk_tag(),
-                viewKey,
-                undefined, // compute input_ids from view_key
-                undefined, // no record inputs
-            );
-            expect(externallySignedRequest.program_id()).to.equal(signedRequest.program_id());
-            expect(externallySignedRequest.function_name()).to.equal(signedRequest.function_name());
-            expect(externallySignedRequest.inputs().length).to.equal(signedRequest.inputs().length);
-            expect(externallySignedRequest.input_ids().length).to.equal(signedRequest.input_ids().length);
-            expect(externallySignedRequest.toString()).to.equal(signedRequest.toString());
-
-            // Build the same request via fromExternallySignedData using pre-computed input_ids (view_key not needed).
-            const externallySignedFromInputIds = ExecutionRequest.fromExternallySignedData(
-                "credits.aleo",
-                "transfer_public",
-                transferPublicInputs,
-                transferPublicInputTypes,
-                signedRequest.signature(),
-                signedRequest.tvk(),
-                signedRequest.signer(),
-                signedRequest.sk_tag(),
-                undefined, // view_key not needed when input_ids provided
-                signedRequest.input_ids(),
-                undefined,
-            );
-            expect(externallySignedFromInputIds.toString()).to.equal(signedRequest.toString());
-        });
-
-        it('Should match ExecutionRequest.sign for transfer_private', async () => {
-            // Use beacon private key - record is owned by beacon address
-            const privateKey = PrivateKey.from_string(beaconPrivateKeyString);
-            const beaconRecord = RecordPlaintext.fromString(recordBeaconOwned);
-            const gamma = beaconRecord.gamma("credits.aleo", "credits", privateKey);
-            const beaconViewKey = ViewKey.from_private_key(PrivateKey.from_string(beaconPrivateKeyString));
-            const externalSigningInputs = await programManager.computeExternalSigningInputs({
-                programName: "credits.aleo",
-                functionName: "transfer_private",
-                inputs: transferPrivateInputs,
-                inputTypes: transferPrivateInputTypes,
-                isRoot: true,
-                checksum: null,
-                viewKey: beaconViewKey,
-            });
-            expect(externalSigningInputs.signer).to.be.ok;
-            expect(externalSigningInputs.skTag).to.be.ok;
-            const signedRequest = ExecutionRequest.sign(
-                privateKey,
-                "credits.aleo",
-                "transfer_private",
-                transferPrivateInputs,
-                transferPrivateInputTypes,
-                undefined,
-                undefined,
-                true,
-            );
-            expect(signedRequest.program_id()).to.equal("credits.aleo");
-            expect(signedRequest.function_name()).to.equal("transfer_private");
-            expect(signedRequest.inputs().length).to.equal(3);
-            expect(externalSigningInputs.requestInputs.length).to.equal(signedRequest.inputs().length);
-
-            const signature = signedRequest.signature();
-            const tvk = signedRequest.tvk();
-            // Use pre-computed input_ids to avoid view_key Option conversion issues across the WASM boundary
-            const externallySignedRequest = ExecutionRequest.fromExternallySignedData(
-                "credits.aleo",
-                "transfer_private",
-                transferPrivateInputs,
-                transferPrivateInputTypes,
-                signature,
-                tvk,
-                signedRequest.signer(),
-                signedRequest.sk_tag(),
-                undefined, // view_key not needed when input_ids provided
-                signedRequest.input_ids(),
-                undefined, // gammas not needed when input_ids provided
-            );
-            expect(externallySignedRequest.program_id()).to.equal(signedRequest.program_id());
-            expect(externallySignedRequest.function_name()).to.equal(signedRequest.function_name());
-            expect(externallySignedRequest.inputs().length).to.equal(signedRequest.inputs().length);
-            expect(externallySignedRequest.input_ids().length).to.equal(signedRequest.input_ids().length);
-            expect(externallySignedRequest.toString()).to.equal(signedRequest.toString());
-        });
-
-        it('Should compute external signing inputs and return ExternalSigningInput shape for private transfer', async () => {
-            const externalSigningInputs = await programManager.computeExternalSigningInputs({
-                programName: "credits.aleo",
-                functionName: "transfer_private",
-                inputs: transferPrivateInputs,
-                inputTypes: transferPrivateInputTypes,
-                isRoot: true,
-                checksum: null,
-            })
-
-            expect(externalSigningInputs).to.be.an("object");
-            expect(externalSigningInputs).to.have.property("functionId");
-            expect(externalSigningInputs).to.have.property("isRoot");
-            expect(externalSigningInputs).to.have.property("requestInputs");
-            expect(externalSigningInputs.functionId).to.be.a("string");
-            expect(externalSigningInputs.functionId.length).to.be.greaterThan(0);
-            expect(externalSigningInputs.functionId).to.match(/field$/);
-            expect(externalSigningInputs.isRoot).to.equal("1field");
-            expect(externalSigningInputs.checksum).to.satisfy((v: unknown) => v === null || v === undefined);
-            expect(externalSigningInputs.requestInputs).to.be.an("array").with.lengthOf(3);
-
-            const [firstInput, secondInput, thirdInput] = externalSigningInputs.requestInputs;
-            expect(firstInput).to.have.property("outputType", "record");
-            expect(firstInput).to.have.property("index", "0field");
-            expect(firstInput).to.have.property("data").that.is.an("array");
-            expect(firstInput.data.length).to.be.greaterThan(0);
-            expect(secondInput).to.have.property("outputType", "private");
-            expect(secondInput).to.have.property("index", "1field");
-            expect(secondInput).to.have.property("data").that.is.an("array");
-            expect(thirdInput).to.have.property("outputType", "private");
-            expect(thirdInput).to.have.property("index", "2field");
-            expect(thirdInput).to.have.property("data").that.is.an("array");
-
-            // Assert each request input's field representation round-trips through Plaintext (fromFields -> toFields -> string).
-            const assertFieldsRoundTripThroughPlaintext = (fieldStrings: string[], expectedString: string) => {
-                const fields = fieldStrings.map((s: string) => Field.fromString(s));
-                const plaintext = Plaintext.fromFields(fields);
-                const backFields = plaintext.toFields();
-                const backStrings = Array.from(backFields).map((f: Field) => f.toString());
-                expect(backStrings).to.deep.equal(fieldStrings);
-                expect(plaintext.toString()).to.equal(expectedString);
-            };
-            // // Assert each request input's field representation round-trips through RecordPlaintext (fromFields -> toFields -> string).
-            // const assertFieldsRoundTripThroughRecordPlaintext = (fieldStrings: string[], expectedString: string) => {
-            //     const fields = fieldStrings.map((s: string) => Field.fromString(s));
-            //     const plaintext = RecordPlaintext.fromFields(fields);
-            //     const backFields = plaintext.toFields();
-            //     const backStrings = Array.from(backFields).map((f: Field) => f.toString());
-            //     expect(backStrings).to.deep.equal(fieldStrings);
-            //     expect(plaintext.toString()).to.equal(expectedString);
-            // };
-            // assertFieldsRoundTripThroughRecordPlaintext(firstInput.data, transferPrivateInputs[0]);
-            assertFieldsRoundTripThroughPlaintext(secondInput.data, transferPrivateInputs[1]);
-            assertFieldsRoundTripThroughPlaintext(thirdInput.data, transferPrivateInputs[2]);
-        });
-    });
 });
