+++ wasm/build.js
@@ -16,9 +16,6 @@ async function buildRollup(input, output) {
     }
 }
 
-
-const isDebugBuild = process.env.BUILD_DEBUG === "1";
-
 async function buildWasm(network) {
     await buildRollup({
         input: {
@@ -27,7 +24,6 @@ async function buildWasm(network) {
         },
         plugins: [
             rust({
-                optimize: isDebugBuild ? { release: false, wasmOpt: false, rustc: false } : undefined,
                 extraArgs: {
                     cargo: [
                         "--no-default-features",
